PR points addressed
