+++ tests/Feature/Http/Web/PortalController/GetIndexTest.php
@@ -41,7 +41,7 @@ class GetIndexTest extends TestCase
             ->assertDontSeeText('You have not made any enquiries yet');
     }
 
-    public function test_redirect_for_authenticated_user_with_incomplete_profile(): void
+    public function test_redirected_to_profile_edit_page_for_authenticated_user_with_incomplete_profile(): void
     {
         $this->authenticateUser(User::factory()->create(['has_completed_profile' => false]));
